+++ CHANGELOG.md
@@ -3658,3 +3658,4 @@
 - [2026-08-16 20:08] refactor(forms): optimize data flow between domain and data layers
 - [2026-08-16 20:10] docs(network): update readme and API contracts for network
 - [2026-08-16 20:32] docs(core): update readme and API contracts for core
+- [2026-08-16 20:45] feat(theme): add async theme stream provider